Create Minimum Cost.cpp
Problem statement:
There are ‘N’ numbers of balls in a room that are placed in a row. You are given an array ‘location’ where location[ i ] denotes the location of the ‘i-th’ ball.

You have to move all the balls at the same location, and it is given that you can move a ball from position [ i ] to

1. position[i] + 2 with cost = 0.

2. position[i] + 1 with cost = 1.

Your task is to find the minimum cost required to move all the balls at the same location.