Create  Two Sum.cpp
Problem statement:
You are given an array of integers 'ARR' of length 'N' and an integer Target. Your task is to return all pairs of elements such that they add up to Target.

Note:

We cannot use the element at a given index twice.
Follow Up:

Try to do this problem in O(N) time complexity.