Create  Check If Linked List Is Palindrome.cpp
Problem statement:
You are given a Singly Linked List of integers. You have to return true if the linked list is palindrome, else return false.



A Linked List is a palindrome if it reads the same from left to right and from right to left.



Example:
The lists (1 -> 2 -> 1), (3 -> 4 -> 4-> 3), and (1) are palindromes, while the lists (1 -> 2 -> 3) and (3 -> 4) are not.