Create  Median of two sorted arrays.cpp
Problem statement:
Given two sorted arrays 'a' and 'b' of size 'n' and 'm' respectively.



Find the median of the two sorted arrays.



Median is defined as the middle value of a sorted list of numbers. In case the length of list is even, median is the average of the two middle elements.



The expected time complexity is O(min(logn, logm)), where 'n' and 'm' are the sizes of arrays 'a' and 'b', respectively, and the expected space complexity is O(1).



Example:
Input: 'a' = [2, 4, 6] and 'b' = [1, 3, 5]

Output: 3.5

Explanation: The array after merging 'a' and 'b' will be { 1, 2, 3, 4, 5, 6 }. Here two medians are 3 and 4. So the median will be the average of 3 and 4, which is 3.5.