Create  Rotate Array ll.cpp
Problem statement:
You are given an array ‘NUM’ containing ‘N’ integers, the direction of rotation ‘D’, and the number of rotations ‘R’.

You have to rotate the given array ‘R’ times. If the direction of rotation ‘D’ has a value equal to ‘f’ then you have to rotate the array forward else if it has a value equal to ‘b’ then rotate the array backward.

For Example :
If N = 3, num = [2, 4, 7], D = ‘f’, R = 1.
You have to rotate the array 1 time in the forward direction. Hence, the output will be [7, 2, 4].