Create  Sorting the stack recursively.cpp
Problem statement:
Ritik is stuck on a problem for the past few days. He has a stack 'S' of integers and his friend told him to sort the stack recursively.

Help Ritik to solve his problem so that he can show it off to his friend.

Note:
1. Looping through the stack is not allowed.
2. You need to return a stack that is sorted in descending order.
For example:
Given stack 'S' = 1 3 2 
The output will be 3 2 1 since it is the sorted order.