Create  Change Case.cpp
Problem statement:
You are given a non-empty string ‘STR’, consisting of both lowercase and uppercase English alphabetical letters. Your task is to convert the characters of the string into the opposite case, i.e. if a character is in lower case then convert it into the upper case and vice-versa.

For Example:
If STR = ‘aBc’,
After converting ‘a’ to ‘A’, ‘B’ to ‘b’, and ‘c’ to ‘C’ we got our final string as ‘AbC’.